Hello there.
You've found me. The big, important commit. The one where I went from some random collection of files into an actual readable repository.

I won't tell you all my secrets. Just know that I've changed a lot of things, soon I'll be buried deep in this project and I'll be very hard to find.

Anyway the changes here are big and important but not hard to follow. There's a bunch of junk that I deleted because I haven't used it in months and the direction of the project has moved away from that stuff anyway. Experiments with CPP and Cython are gone, the earliest expirements I had with python are all gone because they were terrible and sucked and thus not very useful anymore.

I've rebuilt the python directory to have a small number of testing files (which will probably be refactored and made into more helpful files) from the sigma delta estimation testing I was doing, and I have created the one new development file to rule them all. That's the motion_detection.py file sitting all by its lonesome in the root of the python/ directory.

There's now a README that will be updated throughout the lifespan of the project to keep myself in order and keep a line of tasks in development.

That should mostly cover the changes. Enjoy.